+++ Exercise/src/main/java/com/exercise/springdemo/repository/CategoryRepository.java
@@ -3,5 +3,7 @@ package com.exercise.springdemo.repository;
 import com.exercise.springdemo.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+import javax.persistence.criteria.CriteriaBuilder;
+
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
